feat: Move network features to cores_data package
