+++ packages/cores/data/lib/network.dart
@@ -1,0 +1,1 @@
+export 'src/network/index.dart';
