+++ packages/cores/data/lib/src/network/json_type.dart
