+++ packages/cores/data/lib/src/network/provider/dio.dart
@@ -1,6 +1,6 @@
+import 'package:cores_core/exception.dart';
+import 'package:cores_data/src/network/provider/github_access_token.dart';
 import 'package:dio/dio.dart';
-import 'package:flutter_app/core/app_exception.dart';
-import 'package:flutter_app/data/provider/github_access_token.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'dio.g.dart';
